fixes to up_tactic.sh and update_server_and_relaunch.sh

+++ tactic_app/up_tactic.sh
@@ -21,9 +21,15 @@ done
 
 if [ $up_only == "False" ] ; then
   echo "*** removing tactic containers ***"
-  sudo docker ps --filter label="project=tactic" -aq | xargs sudo docker stop | xargs sudo docker rm
+  num=$(sudo docker ps --filter label="project=tactic" -aq | wc -l)
+  if [ $num != "0" ] ; then
+    sudo docker ps --filter label="project=tactic" -aq | xargs sudo docker stop | xargs sudo docker rm
+  fi
   echo "*** removing aux containers ***"
-  sudo docker ps --filter label="project=tactic_aux" -aq | xargs sudo docker stop | xargs sudo docker rm
+  num=$(sudo docker ps --filter label="project=tactic_aux" -aq | wc -l)
+  if [ $num != "0" ] ; then
+    sudo docker ps --filter label="project=tactic_aux" -aq | xargs sudo docker stop | xargs sudo docker rm
+  fi
 fi
 
 sudo docker compose --env-file $env_file --profile start_project up --detach
